refactor(protocol): 更新协议格式和消息类型编码
- 将操作类型字段从1字节扩展为2字节，协议头总长度从10字节增加到15字节
- 修改ChecksumHandler中的头部字段读取逻辑，使用索引定位而非顺序读取
- 更新GameMessageDecoder和GameMessageEncoder以支持2字节操作类型字段
- 调整GameMsgType枚举值，将服务器到客户端消息类型值增加到5位数
- 修改协议常量定义以反映新的协议头字段长度配置

+++ src/main/java/com/zunf/tankbattletcpserver/handler/netty/GameMessageEncoder.java
@@ -25,7 +25,7 @@ public class GameMessageEncoder extends MessageToMessageEncoder<GameMessage> {
         ByteBuf buf = alloc.buffer(ProtocolConstant.HEADER_TOTAL_LENGTH + bodyLength);
 
         // 写头部
-        buf.writeByte(msg.getMsgType().getCode());
+        buf.writeShort(msg.getMsgType().getCode());
         buf.writeByte(msg.getVersion());
         buf.writeInt(msg.getRequestId());
         buf.writeInt(bodyLength);
